Ensure edits to `theme.json` update Tailwind config during `watch` (#79) There were two problems here:
* Opening the `theme.json` file with `require` now caches its contents, and those contents aren’t updated when the file changes.
* Tailwind no longer executes the plugin except on startup and when rebuilding the configuration after changes to `tailwind.config.js`.

To overcome the caching issue, `require` is no longer used to load the `theme.json` file; that logic has been moved to the `@_tw/themejson` package (in a backwards-compatible manner).

To force Tailwind to rebuild the configuration, `chokidar-cli` is used to watch the `theme.json` file and to trigger `nodetouch` on the `tailwind.config.js` file when `theme.json` changes.

Depending on how things change post-Oxide, this may need to be revisited.